+++ testbench/TbAxi4/TestCtrl_e.vhd
@@ -65,6 +65,7 @@ entity TestCtrl is
   ) ;
   port (
     -- Global Signal Interface
+    Clk               : In    std_logic := '0' ;
     nReset            : In    std_logic ;
 
     -- Transaction Interfaces